refactor: renamed shared directory in frontend/components to _shared
