+++ frontend/src/components/_shared/NavBar.jsx
