+++ frontend/src/components/_shared/NumberField.jsx
