+++ frontend/src/components/_shared/ToastSettings.jsx
