+++ frontend/src/components/groupsPage/GroupCard.jsx
@@ -1,9 +1,9 @@
 import { Spacer, HStack, VStack, Box } from '@chakra-ui/react';
 import { ViewIcon } from "@chakra-ui/icons";
 import { useParams } from 'react-router';
-import StudentsPreviewTable from "../shared/StudentsPreviewTable";
-import NavButton from "../shared/NavButton";
-import PageHeader from "../shared/PageHeader";
+import StudentsPreviewTable from "../_shared/StudentsPreviewTable";
+import NavButton from "../_shared/NavButton";
+import PageHeader from "../_shared/PageHeader";
 
 const GroupCard = ({groupData, numberOfGroups}) => {
     const {
