+++ frontend/src/components/importPage/AddStudentModal.jsx
@@ -16,10 +16,10 @@ import {
 } from "@chakra-ui/react";
 import React, { useState } from "react";
 import { AddIcon } from "@chakra-ui/icons";
-import TextField from "../shared/TextField";
-import DropdownDynamic from "../shared/DropdownDynamic";
-import NumberField from "../shared/NumberField";
-import getToastSettings from "../shared/ToastSettings";
+import TextField from "../_shared/TextField";
+import DropdownDynamic from "../_shared/DropdownDynamic";
+import NumberField from "../_shared/NumberField";
+import getToastSettings from "../_shared/ToastSettings";
 import {MockAuth} from "../../helpers/mockAuth";
 import {useAuth0} from "@auth0/auth0-react";
