+++ frontend/src/components/unitHomePage/DeleteModal.jsx
@@ -7,10 +7,10 @@ import {
     ModalCloseButton,
     useToast
 } from '@chakra-ui/react'
-import ModalFooterButtonPair from "../shared/ModalFooterButtonPair";
+import ModalFooterButtonPair from "../_shared/ModalFooterButtonPair";
 import {MockAuth} from "../../helpers/mockAuth";
 import {useAuth0} from "@auth0/auth0-react";
-import getToastSettings from "../shared/ToastSettings";
+import getToastSettings from "../_shared/ToastSettings";
 
 
 const DeleteModal = ({
