+++ frontend/src/pages/GroupAnalytics.jsx
@@ -15,8 +15,8 @@ import {
 } from 'chart.js';
 
 import { MockAuth } from '../helpers/mockAuth';
-import PageHeader from "../components/shared/PageHeader";
-import NavButton from "../components/shared/NavButton";
+import PageHeader from "../components/_shared/PageHeader";
+import NavButton from "../components/_shared/NavButton";
 import AnalyticsCard from "../components/analyticsPage/AnalyticsCard";
 
 ChartJS.register(ArcElement, Tooltip, Legend, LinearScale, CategoryScale, BarController, BarElement);
