+++ frontend/src/pages/Groups.jsx
@@ -21,10 +21,10 @@ import {
 
 import GroupCard from '../components/groupsPage/GroupCard';
 import { MockAuth } from '../helpers/mockAuth';
-import NavButton from "../components/shared/NavButton";
-import ToggleButtonGroup from "../components/shared/ToggleButtonGroup";
-import PageHeader from "../components/shared/PageHeader";
-import getToastSettings from '../components/shared/ToastSettings';
+import NavButton from "../components/_shared/NavButton";
+import ToggleButtonGroup from "../components/_shared/ToggleButtonGroup";
+import PageHeader from "../components/_shared/PageHeader";
+import getToastSettings from '../components/_shared/ToastSettings';
 
 function Groups() {
     const [groups, setGroups] = useState([]);
